restructure workChatroom files, fix post mic button press send error
