+++ client/src/pages/chatrooms/workChat/Messaging.js
@@ -1,10 +1,8 @@
 import React, { useState, useRef, useEffect } from "react"
-import microphoneButton from "../assets/images/microphone-button.png"
-import sendButton from "../assets/images/send-button.png"
-import useRecorder from "../hooks/useRecorder"
+import useRecorder from "../../../hooks/useRecorder"
 import axios from "axios"
 
-import { firestore, auth } from "../services/firebase"
+import { firestore, auth } from "../../../services/firebase"
 import SendMessage from "./SendMessage"
 // import microphoneButton from "../assets/images/microphone-button.png"
 // import sendButton from "../assets/images/send-button.png"
@@ -49,7 +47,7 @@ const Messaging = ({ title, dbCollection }) => {
 
   useEffect(() => {
     firestore
-      .collection("messages")
+      .collection("workMessages")
       // .collection(`{dbCollection}`)
       .orderBy("createdAt")
       .limit(50)
@@ -72,9 +70,9 @@ const Messaging = ({ title, dbCollection }) => {
   var item = items[Math.floor(Math.random() * items.length)]
 
   return (
-    <div className=' w-full h-screen '>
+    <div className='w-full h-screen '>
       <div className='pt-4'>
-        <div className='bg-white px-4 text-primary uppercase text-4xl grid place-items-left rounded-lg py-1'>
+        <div className='grid px-4 py-1 text-4xl uppercase bg-white rounded-lg text-primary place-items-left'>
           {title}
         </div>
       </div>
@@ -82,7 +80,7 @@ const Messaging = ({ title, dbCollection }) => {
       <div className='px-5 py-6 '>
         <div
           style={{ height: "65vh" }}
-          className='rounded-lg bg-slate-200 p-4 overflow-y-scroll flex flex-col h-128 w-full'
+          className='flex flex-col w-full p-4 overflow-y-scroll rounded-lg bg-slate-200 h-128'
         >
           {messages.map(({ id, text, uid }) => (
             <div>
@@ -101,7 +99,7 @@ const Messaging = ({ title, dbCollection }) => {
                     : "received rounded-br-3xl bg-[#353df0] float-left"
                 }`}
               >
-                <p className='font-medium text-md break-words'>{text}</p>
+                <p className='font-medium break-words text-md'>{text}</p>
               </div>
             </div>
           ))}
@@ -115,11 +113,12 @@ const Messaging = ({ title, dbCollection }) => {
         scroll={scroll}
         setStartedRecording={setStartedRecording}
         startRecord={startRecord}
+        isRecording={isRecording}
         // messageDbCollection={dbCollection}
       />
       {/* <audio src={audioURL} controls />
-      <div className='mx-5 py-6'>Transcript: {text}</div>
-      <div className='py-5 flex flex-row px-5 '>
+      <div className='py-6 mx-5'>Transcript: {text}</div>
+      <div className='flex flex-row px-5 py-5 '>
         <button
           onClick={(e) => {
             setStartedRecording(true)
@@ -141,7 +140,7 @@ const Messaging = ({ title, dbCollection }) => {
               ? "Press the microphone to stop recording once you're done"
               : "Press the microphone vent out"
           }
-          className='px-3  border-2 border-primary  w-full rounded-xl'
+          className='w-full px-3 border-2 border-primary rounded-xl'
         ></input>
         <button>
           <img className='cursor-pointer' src={sendButton} />
